refactor packages move config upgrade version in requirements
