+++ fiwareglancesync/app/config.py
@@ -26,6 +26,7 @@ import os
 
 __author__ = 'fla'
 
+
 # Statement for enabling the development environment
 DEBUG = True
