+++ setup.py
@@ -23,7 +23,7 @@
 # contact with opensource@tid.es
 #
 from setuptools import setup, find_packages
-from glancesync.settings.glancesync_config import __version__
+from fiwareglancesync.settings.glancesync_config import __version__
 from pip.req import parse_requirements
 from os.path import join as pjoin
 
@@ -32,21 +32,19 @@ install_reqs = parse_requirements("requirements.txt", session=False)
 # > requirements_list is a list of requirement; e.g. ['requests==2.6.0', 'Fabric==1.8.3']
 requirements_list = [str(ir.req) for ir in install_reqs]
 
+
 setup(
   name='fiware-glancesync',
-  packages=find_packages(exclude=['*tests*']),
   install_requires=requirements_list,
-  data_files=[('.', ['glancesync/settings/glancesync.conf'])],
-  package_data={
-    'glancesync': ['glancesync.conf']
-  },
-  include_package_data=True,
+  data_files=[('.', ['fiwareglancesync/settings/glancesync.conf']), ('.',
+                                                                     ['fiwareglancesync/app/config.py'])],
+  packages=find_packages(),
   version=__version__,
   description='Tool to synchronise images from a master region to other regions',
   author='Fernando Lopez Aguilar',
   author_email='fernando.lopezaguilar@telefonica.com, e.fiware.tid@telefonica.com',
   license='Apache 2.0',
-  scripts=[pjoin('sync.py')],
+  scripts=[pjoin('fiwareglancesync/run.py'),pjoin('fiwareglancesync/sync.py')],
   url='https://github.com/telefonicaid/fiware-glancesync',
   download_url='https://github.com/telefonicaid/fiware-glancesync/tarball/v%s' % __version__,
   keywords=['fiware', 'glancesync', 'glance',  'images', 'cloud'],
